refactor(preview): share native document model
