+++ src/ui/settings/tests.rs
@@ -80,7 +80,7 @@ fn available_notes_are_shown_only_for_a_newer_release() {
             version: "1.0.0".to_owned(),
             url: "https://example.test/release".to_owned(),
             notes: "Changes".to_owned(),
-            note_blocks: vec![crate::services::ReleaseNoteBlock::Paragraph(
+            note_blocks: vec![crate::services::DocumentBlock::Paragraph(
                 "Changes".to_owned(),
             )],
             kind: BuildKind::Stable,
